chore(scripts): remove releaser v2 flow and drop v1 naming
Make the interactive release wizard the only release path.

- Delete .github/workflows/tag-and-release.yaml and scripts/releaser/v2.
- Move scripts/releaser/v1 into scripts/releaser (package main) and drop
  the --legacy flag and v2 subcommands from main.go.
- Update scripts/release.sh to invoke the wizard directly.
