+++ scripts/releaser/git.go
@@ -1,4 +1,4 @@
-package v1
+package main
 
 import (
 	"errors"
